+++ tests/testthat/test-read_vcf.R
@@ -4,7 +4,8 @@ test_that("Test that VCFs can be properly read in.", {
     ### Read in original VCF
     sumstats_dt  = read_vcf(path = vcf_path) 
     sumstats_dt <- standardise_sumstats_column_headers_crossplatform(sumstats_dt = sumstats_dt, 
-                                                      path=vcf_path)[["sumstats_dt"]]
+                                                      path=vcf_path,
+                                                      mapping_file=sumstatsColHeaders)[["sumstats_dt"]]
     ### Write this VCF and then read it in again
     tmp_vcf <- tempfile(fileext='.vcf.gz')
     write_sumstats(sumstats_dt = sumstats_dt, 
@@ -13,7 +14,8 @@ test_that("Test that VCFs can be properly read in.", {
                    write_vcf = TRUE)
     sumstats_dt2 <- read_vcf(path = tmp_vcf) 
     sumstats_dt2 <- standardise_sumstats_column_headers_crossplatform(sumstats_dt = sumstats_dt, 
-                                                                         path=vcf_path)[["sumstats_dt"]]
+                                                                         path=vcf_path,
+                                                                      mapping_file=sumstatsColHeaders)[["sumstats_dt"]]
     expect_equal(sumstats_dt, sumstats_dt2)
 })